refactor : d3코드 컴포넌트에서 분리 #102
